% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{get_bicycle_routes}
\alias{get_bicycle_routes}
\title{Turku bicycle routes}
\source{
See
  \href{https://www.avoindata.fi/data/fi/dataset/turun-kaupungin-katujen-kunnossapitorajapinta}{avoindata.fi}
   for additional information. See
  \href{https://github.com/City-of-Helsinki/aura/wiki/API}{City of Helsinki documentation}
  for more information on the API.
}
\usage{
get_bicycle_routes(dataset = NULL, ...)
}
\arguments{
\item{choice}{Input "paa", "lahi" or NULL (default)}
}
\value{
spatial object
}
\description{
Download and visualize bicycle routes in Turku
}
\examples{
bicycle <- get_bicycle_routes(dataset = "paa")
}
\author{
Pyry Kantanen
}
