% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{get_citybike_stats}
\alias{get_citybike_stats}
\title{Turku city bike data}
\source{
Source: Turku region public transport's transit and timetable data.
The administrator of data is Turku region public transport.
Dataset is downloaded from http://data.foli.fi/ using the license
Creative Commons Attribution 4.0 International (CC BY 4.0).

See \href{http://data.foli.fi/doc/index-en}{TSJL - transit API}
and \href{http://data.foli.fi/doc/citybike/v0/index}{FÖLLÄRI - Kaupunkipyörät}
for additional information.
}
\usage{
get_citybike_stats(format = NULL)
}
\arguments{
\item{format}{Input NULL (default), "json" or "geojson".}
}
\value{
data.frame or sf object
}
\description{
Download city bike data in Turku
}
\examples{
follari_stations <- get_citybike_stats(format = "geojson")
}
\author{
Pyry Kantanen
}
