% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{get_maintenance_locations}
\alias{get_maintenance_locations}
\title{Turku street maintenance API}
\source{
See
  \href{https://www.avoindata.fi/data/fi/dataset/turun-kaupungin-katujen-kunnossapitorajapinta}{avoindata.fi}
   for additional information. See
  \href{https://github.com/City-of-Helsinki/aura/wiki/API}{City of Helsinki documentation}
  for more information on the API.
}
\usage{
get_maintenance_locations(id = NULL, ...)
}
\arguments{
\item{id}{Vehicle ID as integer}

\item{...}{List queries, for example vehicle id as integer,
\code{history} (integer), \code{since} (integer), \code{limit} (integer),
\code{temporal_resolution} (integer)}
}
\value{
data.frame
}
\description{
Download and list where maintenance vehicles are located and
  what they are doing.
}
\details{
Possible events for vehicles: \itemize{ \item{"kv"} {Work done on
  bicycle and pedestrian lanes} \item{"au"} {Snow removal (auraus)}
  \item{"su"} {De-icing with salt (suolaus)} \item{"hi"} {Spreading sand
  (hiekoitus)} \item{"nt"} {Mowing (niitto)} \item{"ln"} {Levelling down high
  spots (Lanaus)} \item{"hs"} {Planing (höyläys)} \item{"pe"} {Street washing
  (kadunpesu)} \item{"ps"} {Dust binding (Pölynsidonta)} \item{"hn"} {Sand
  removal (hiekannosto)} \item{"hj"} {Brushing (harjaus)} \item{"pn"}
  {Coating (pinnoitus, liittyy kesähoitoreitteihin} }
}
\examples{
## vehicle 8, 10 last instances
vehicles <- get_maintenance_locations(id = 8, history = 10)
}
\author{
Pyry Kantanen
}
