% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{get_traffic_count}
\alias{get_traffic_count}
\title{Turku traffic counter data}
\source{
Source: Turku region public transport's transit and timetable data.
The administrator of data is Turku region public transport.
Dataset is downloaded from http://data.foli.fi/ using the license
Creative Commons Attribution 4.0 International (CC BY 4.0).

See \href{http://data.foli.fi/doc/index-en}{TSJL - transit API}
and \href{http://data.foli.fi/doc/citybike/v0/index}{FÖLLÄRI - Kaupunkipyörät}
for additional information.
}
\usage{
get_traffic_count(
  data.url = NULL,
  use.sf = FALSE,
  vehicle.type = "J",
  direction = "K"
)
}
\arguments{
\item{data.url}{Option to input own url}

\item{use.sf}{Return an sf object. Default is FALSE}

\item{vehicle.type}{Select data from certain traffic sensors: "A" (cars),
"J" (pedestrians) or "P" (bicycles)}

\item{direction}{Traffic to ("K") or from ("P") city centre}
}
\value{
data.frame or sf object
}
\description{
Download data from traffic data counters
}
\examples{
follari_stations <- get_traffic_count(vehicle.type = "A", direction = "K")
}
\author{
Pyry Kantanen
}
