% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{get_turku_addresses}
\alias{get_turku_addresses}
\title{List Turku region addresses}
\source{
See \href{https://www.avoindata.fi/data/fi/dataset/turun-seudun-osoitteet}{avoindata.fi}
for additional information.
}
\usage{
get_turku_addresses(city = NULL, to.sf = FALSE, fix.data = TRUE)
}
\arguments{
\item{city}{Choose 1 or more cities. Default is NULL, returning the whole
dataset}

\item{to.sf}{Turn list into an sf object. Default is FALSE}

\item{fix.data}{Quick and dirty fix for erroneous input in data. Default is
TRUE.}
}
\value{
list or sf object
}
\description{
Download and list addresses in Turku region
}
\details{
Dataset contains addresses for Turku, Kaarina, Aura, Lieto,
Marttila, Paimio, Sauvo, Rusko, Raisio, Masku, Nousiainen, Mynämäki and Naantali.

Coordinates are given in EPSG:3877 (ETRS-GK23) format
}
\examples{
addresses <- get_turku_addresses(city = c("Turku", "Kaarina"), to.sf = FALSE)
}
\author{
Pyry Kantanen
}
