% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{get_airmonitoring}
\alias{get_airmonitoring}
\title{Access Helsinki region Service Map API}
\source{
API Location: https://api.turku.fi/airmonitoring/v1/

API documentation: https://api.turku.fi/airmonitoring/v1/
}
\usage{
get_airmonitoring(station.id = NULL, to.sf = FALSE, timeseries = FALSE)
}
\arguments{
\item{station.id}{NULL (default) returns information from all stations, inputing
one or more ID's returns information for given stations. Valid stations are:
"kauppatori", "ruissalo", "raisio", "naantali", "kaarina", "kaanaa" and "parainen".}

\item{to.sf}{FALSE (default) returns a data.table, TRUE returns an sf object
for easy visualization on map}

\item{timeseries}{FALSE (default) returns the last observation for station(s),
TRUE returns historical data}
}
\value{
a data.frame or an sf object
}
\description{
Access the new Helsinki region Service Map (Paakaupunkiseudun Palvelukartta)
http://dev.hel.fi/servicemap/ data through the API: http://api.hel.fi/servicemap/v2/.
For more API documentation and license information see the API link.
}
\examples{
kauppatori <- get_airmonitoring(station.id="kauppatori")

}
\author{
Pyry Kantanen
}
