% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{get_turku_energia_art}
\alias{get_turku_energia_art}
\title{Access Turku Region Linked Events API}
\source{
See \href{https://www.avoindata.fi/data/fi/dataset/turku-energian-ymparistotaideteokset}{avoindata.fi}
for additional information.
}
\usage{
get_turku_energia_art(
  data.url = "https://dev.turku.fi/datasets/turku-energia-ymparistotaideteokset.csv"
)
}
\arguments{
\item{data.url}{Dataset url}
}
\value{
sf object
}
\description{
Easy access to Turku Region Linked Events API
}
\examples{
art_sf <- get_turku_energia_art()

}
\author{
Pyry Kantanen
}
